CRA-FE-3 add cocktail card to list [SD,SC]
