+++ Frontend/Android/app/src/main/res/layout/activity_main.xml
@@ -17,7 +17,7 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <include
-        layout="@layout/cocktail_list_view_card"
+        layout="@layout/cocktail_list_view"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         app:layout_constraintBottom_toBottomOf="parent"
