+++ Frontend/Android/app/src/main/res/layout/cocktail_list_view.xml
@@ -8,7 +8,7 @@
         android:layout_height="match_parent">
 
         <LinearLayout
-            android:id="@+id/cocktail_list_view"
+            android:id="@+id/cocktail_list"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:orientation="vertical" />
